fix: move sarEnable to separate file to fix m5stamp_c3 build tag mismatch
The sarEnable() method references regI2C and analog config constants
defined in machine_esp32xx_adc.go, which excludes m5stamp_c3. Moving it
to a new file with the matching build constraint prevents undefined
symbol errors for that target.
